# cedar-repo-server-application/src/main/java/org/metadatacenter/cedar/repo/resources/TemplateInstancesResource.java
package org.metadatacenter.cedar.repo.resources;

import com.codahale.metrics.annotation.Timed;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import com.fasterxml.jackson.databind.JsonNode;
import com.mongodb.MongoException;
import org.metadatacenter.config.CedarConfig;
import org.metadatacenter.exception.CedarException;
import org.metadatacenter.id.CedarTemplateInstanceId;
import org.metadatacenter.model.CedarResourceType;
import org.metadatacenter.rest.context.CedarRequestContext;
import org.metadatacenter.server.service.TemplateInstanceService;
import org.metadatacenter.util.http.CedarResponse;
import org.metadatacenter.util.json.JsonUtils;

import jakarta.ws.rs.GET;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.PathParam;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;

import static org.metadatacenter.constant.CedarPathParameters.PP_ID;
import static org.metadatacenter.rest.assertion.GenericAssertions.LoggedIn;
import static org.metadatacenter.server.security.model.auth.CedarPermission.TEMPLATE_INSTANCE_READ;

@Path("/template-instances")
@Produces(MediaType.APPLICATION_JSON)
@Tag(name = "Template instances")
@SecurityRequirement(name = "api_key")
public class TemplateInstancesResource extends AbstractRepoResource {

  private final TemplateInstanceService<String, JsonNode> templateInstanceService;

  public TemplateInstancesResource(CedarConfig cedarConfig, TemplateInstanceService<String, JsonNode> templateInstanceService) {
    super(cedarConfig);
    this.templateInstanceService = templateInstanceService;
  }

  @GET
  @Timed
  @Path("/{id}")
  @Operation(summary = "Resolve a template instance identifier",
      description = "Return the template instance a CEDAR identifier names. This is where an artifact's `@id` "
          + "dereferences to, so the path segment is the bare identifier that ends that IRI rather "
          + "than the IRI itself. Reading is checked twice: the caller needs the template instance read "
          + "permission, and then read access to this particular artifact as the workspace records "
          + "it. Mongo's internal `_id` is removed before the artifact is returned.")
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The stored template instance"),
      @ApiResponse(responseCode = "401",
          description = "The request carries no valid credentials, or the caller has no read access "
              + "to this artifact. The second case is a permission failure reported as 401 rather "
              + "than 403."),
      @ApiResponse(responseCode = "403", description = "The caller lacks the template instance read permission"),
      @ApiResponse(responseCode = "404",
          description = "The workspace knows the identifier but the repository holds no such artifact"),
      @ApiResponse(responseCode = "500",
          description = "Internal server error. Also returned for an identifier the workspace does "
              + "not know at all, which is reported as a server error rather than as a 404.")
  })
  public Response findTemplateInstance(
      @Parameter(description = "The bare identifier ending the artifact's IRI, not the whole IRI. "
          + "Example: 8bc64ab5-df6b-48c8-8c61-6c016245918e", required = true)
      @PathParam(PP_ID) String id) throws CedarException {

    CedarRequestContext c = buildRequestContext();

    c.must(c.user()).be(LoggedIn);
    c.must(c.user()).have(TEMPLATE_INSTANCE_READ);

    String templateInstanceId = linkedDataUtil.getLinkedDataId(CedarResourceType.INSTANCE, id);
    CedarTemplateInstanceId tid = CedarTemplateInstanceId.build(templateInstanceId);

    if (userHasNoReadAccessToArtifact(c, tid)) {
      return CedarResponse.unauthorized().build();
    }

    try {
      JsonNode templateInstance = templateInstanceService.findTemplateInstance(templateInstanceId);
      if (templateInstance != null) {
        // Remove autogenerated _id field to avoid exposing it
        templateInstance = JsonUtils.removeField(templateInstance, "_id");
        return Response.ok().entity(templateInstance).build();
      }
      return CedarResponse.notFound().build();
    } catch (MongoException e) {
      throw e;
    } catch (Exception e) {
      return CedarResponse.internalServerError().exception(e).build();
    }
  }

}